move things around, big refactor in to components and pages
