+++ src/components/ChartEntry/ChartEntry.js
