+++ src/components/ChartEntry/ChartEntry.module.scss
@@ -1,4 +1,4 @@
-@import "../rem-calc";
+@import "../../rem-calc";
 
 .chartEntry {
   display: block;
