+++ src/components/Header/Header.module.scss
@@ -1,4 +1,4 @@
-@import "../rem-calc";
+@import "../../rem-calc";
 
 .dropdown {
   position: relative;
