+++ src/components/LineChart/LineChart.module.scss
@@ -1,4 +1,4 @@
-@import "../rem-calc";
+@import "../../rem-calc";
 
 .lineChartContainer {
   display: block;
