+++ src/components/LineChart/colors.json
