+++ src/components/SongList/SongList.module.scss
@@ -1,4 +1,4 @@
-@import "../rem-calc";
+@import "../../rem-calc";
 
 .songList {
   display: block;
