+++ src/index.js
@@ -2,8 +2,8 @@ import React from "react";
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import styles from "./App.module.scss";
-import Header from "./Header/Header";
-import Song from "./Song/Song";
+import Header from "./components/Header/Header";
+import Song from "./pages/Song/Song";
 import "./index.css";
 import ArtistDetailsPage from "./pages/ArtistDetails/ArtistDetailsPage";
 import SearchResultPage from "./pages/SearchResultsPage/SearchResultsPage";
