+++ src/pages/Song/Song.module.scss
@@ -1,4 +1,4 @@
-@import "../rem-calc";
+@import "../../rem-calc";
 
 .songName {
   font-size: rem-calc(16);
